Added basic shadow mapping

+++ Grapple/src/Grapple/Platform/OpenGL/OpenGLTexture.cpp
@@ -115,7 +115,7 @@ namespace Grapple
 
 	void* OpenGLTexture::GetRendererId() const
 	{
-		return (void*)m_Id;
+		return (void*)(size_t)m_Id;
 	}
 
 	void OpenGLTexture::SetFiltering(TextureFiltering filtering)
